:white_check_mark: Separate test classes
