+++ .github/workflows/python-app.yml
@@ -25,7 +25,7 @@ jobs:
           if [ -f requirements.txt ]; then pip install -r requirements.txt; fi
       - name: Test with unittest
         run: |
-          coverage run -m unittest discover -v -s tests -p "*_test.py" -t src && coverage report -m && coverage xml
+          coverage run -m unittest discover -v -s tests -p "test_*.py" -t src && coverage report -m && coverage xml
       - name: Upload coverage to Codecov
         uses: codecov/codecov-action@v1
         with:
